Replace serializable with parcelable and BigInteger with Double

+++ app/src/main/java/com/weatherxm/data/ApiModels.kt
@@ -9,7 +9,6 @@ import com.weatherxm.ui.common.DeviceRelation
 import com.weatherxm.ui.common.UIDevice
 import kotlinx.parcelize.Parcelize
 import timber.log.Timber
-import java.math.BigInteger
 import java.time.LocalDate
 import java.time.ZonedDateTime
 
@@ -524,11 +523,11 @@ data class QoDErrorAffects(
 data class WalletRewards(
     val proof: List<String>?,
     @Json(name = "cumulative_amount")
-    val cumulativeAmount: BigInteger?,
+    val cumulativeAmount: Double?,
     val cycle: Int?,
-    val available: BigInteger?,
+    val available: Double?,
     @Json(name = "total_claimed")
-    val totalClaimed: BigInteger?,
+    val totalClaimed: Double?,
 ) : Parcelable
 
 enum class DeviceProfile {
